glusterfs: update for new API
New versions of Glusterfs have an enhanced set of APIs that returns
additional 'struct stat' structures to enable advanced caching of
attributes.  Applications using gfapi would need to adapt to the
newer interfaces to compile against release-6 APIs.

This should fix compilation with v6.0 of the API.
